+++ yash-env/src/system/shared.rs
@@ -235,8 +235,8 @@ impl SharedSystem {
 
     /// Waits for some signals to be delivered to this process.
     ///
-    /// Before calling this function, you need to [set signal
-    /// handling](Self::set_signal_handling) to `Catch`. Without doing so, this
+    /// Before calling this function, you need to [set the signal
+    /// disposition](Self::set_disposition) to `Catch`. Without doing so, this
     /// function cannot detect the receipt of the signals.
     ///
     /// Returns an array of signals that were caught.
@@ -263,8 +263,8 @@ impl SharedSystem {
 
     /// Waits for a signal to be delivered to this process.
     ///
-    /// Before calling this function, you need to [set signal
-    /// handling](Self::set_signal_handling) to `Catch`.
+    /// Before calling this function, you need to [set the signal
+    /// disposition](Self::set_disposition) to `Catch`.
     /// Without doing so, this function cannot detect the receipt of the signal.
     ///
     /// If this `SharedSystem` is part of an [`Env`], you should call
@@ -709,7 +709,7 @@ impl SignalSystem for &SharedSystem {
         System::signal_number_from_name(*self, name)
     }
 
-    fn set_signal_handling(
+    fn set_disposition(
         &mut self,
         signal: signal::Number,
         disposition: Disposition,
@@ -730,7 +730,7 @@ impl SignalSystem for SharedSystem {
     }
 
     #[inline]
-    fn set_signal_handling(
+    fn set_disposition(
         &mut self,
         signal: signal::Number,
         disposition: Disposition,
@@ -918,15 +918,9 @@ mod tests {
         let process_id = system.process_id;
         let state = Rc::clone(&system.state);
         let mut system = SharedSystem::new(Box::new(system));
-        system
-            .set_signal_handling(SIGCHLD, Disposition::Catch)
-            .unwrap();
-        system
-            .set_signal_handling(SIGINT, Disposition::Catch)
-            .unwrap();
-        system
-            .set_signal_handling(SIGUSR1, Disposition::Catch)
-            .unwrap();
+        system.set_disposition(SIGCHLD, Disposition::Catch).unwrap();
+        system.set_disposition(SIGINT, Disposition::Catch).unwrap();
+        system.set_disposition(SIGUSR1, Disposition::Catch).unwrap();
 
         let mut context = Context::from_waker(noop_waker_ref());
         let mut future = Box::pin(system.wait_for_signals());
@@ -960,9 +954,7 @@ mod tests {
         let process_id = system.process_id;
         let state = Rc::clone(&system.state);
         let mut system = SharedSystem::new(Box::new(system));
-        system
-            .set_signal_handling(SIGCHLD, Disposition::Catch)
-            .unwrap();
+        system.set_disposition(SIGCHLD, Disposition::Catch).unwrap();
 
         let mut context = Context::from_waker(noop_waker_ref());
         let mut future = Box::pin(system.wait_for_signal(SIGCHLD));
@@ -989,12 +981,8 @@ mod tests {
         let process_id = system.process_id;
         let state = Rc::clone(&system.state);
         let mut system = SharedSystem::new(Box::new(system));
-        system
-            .set_signal_handling(SIGINT, Disposition::Catch)
-            .unwrap();
-        system
-            .set_signal_handling(SIGTERM, Disposition::Catch)
-            .unwrap();
+        system.set_disposition(SIGINT, Disposition::Catch).unwrap();
+        system.set_disposition(SIGTERM, Disposition::Catch).unwrap();
 
         let mut context = Context::from_waker(noop_waker_ref());
         let mut future = Box::pin(system.wait_for_signal(SIGINT));
@@ -1019,12 +1007,8 @@ mod tests {
         let process_id = system.process_id;
         let state = Rc::clone(&system.state);
         let mut system = SharedSystem::new(Box::new(system));
-        system
-            .set_signal_handling(SIGINT, Disposition::Catch)
-            .unwrap();
-        system
-            .set_signal_handling(SIGTERM, Disposition::Catch)
-            .unwrap();
+        system.set_disposition(SIGINT, Disposition::Catch).unwrap();
+        system.set_disposition(SIGTERM, Disposition::Catch).unwrap();
 
         {
             let mut state = state.borrow_mut();
@@ -1052,7 +1036,7 @@ mod tests {
         let mut system_3 = system_1.clone();
         let (reader, writer) = system_1.pipe().unwrap();
         system_2
-            .set_signal_handling(SIGCHLD, Disposition::Catch)
+            .set_disposition(SIGCHLD, Disposition::Catch)
             .unwrap();
 
         let mut buffer = [0];